adapter-proxy/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
httpx>=0.28,<1
pyyaml>=6.0,<7
jinja2>=3.1,<4
